+++ common-client/src/main/java/de/maxhenkel/voicechat/plugins/impl/VoicechatClientApiImpl.java
@@ -6,11 +6,13 @@ import de.maxhenkel.voicechat.api.audiochannel.ClientEntityAudioChannel;
 import de.maxhenkel.voicechat.api.audiochannel.ClientLocationalAudioChannel;
 import de.maxhenkel.voicechat.api.audiochannel.ClientStaticAudioChannel;
 import de.maxhenkel.voicechat.api.config.ConfigAccessor;
+import de.maxhenkel.voicechat.api.config.VolumeConfigAccessor;
 import de.maxhenkel.voicechat.intercompatibility.ClientCompatibilityManager;
 import de.maxhenkel.voicechat.plugins.impl.audiochannel.ClientEntityAudioChannelImpl;
 import de.maxhenkel.voicechat.plugins.impl.audiochannel.ClientLocationalAudioChannelImpl;
 import de.maxhenkel.voicechat.plugins.impl.audiochannel.ClientStaticAudioChannelImpl;
 import de.maxhenkel.voicechat.plugins.impl.config.ConfigAccessorImpl;
+import de.maxhenkel.voicechat.plugins.impl.config.VolumeConfigAccessorImpl;
 import de.maxhenkel.voicechat.voice.client.*;
 import de.maxhenkel.voicechat.voice.common.ClientGroup;
 
@@ -137,6 +139,11 @@ public class VoicechatClientApiImpl extends VoicechatApiImpl implements Voicecha
     @Override
     public ConfigAccessor getClientConfig() {
         return new ConfigAccessorImpl(VoicechatClient.CLIENT_CONFIG.disabled.getConfig());
+    }
+
+    @Override
+    public VolumeConfigAccessor getVolumeConfig() {
+        return VolumeConfigAccessorImpl.INSTANCE;
     }
 
     @Override