+++ forge/changelog.md
@@ -1,3 +1,1 @@
-- Updated Japanese translation
-- Fixed servers not accepting connections from clients without the mod
-- Fixed MTU size not considered when reading voice chat packets
+- Added volume config accessor to the API